Delete `New*SetFromSlice()` and `NewSetWith()` APIs
These are no longer needed now that both `NewSet()` and
`NewThreadUnsafeSet()` accept optional vals.

Removing them simplifies the API surface, so that developers aren't left
scratching their heads wondering why there's both a `NewSet(vals ...T)`
and a `NewSetFromSlice()` that do the same thing.

Previously they couldn't be removed due to backwards compatibility
concerns, but with the `go 1.18` drop of generics, this library will be
cutting a new `v2` release, so a perfect time to cleanup this confusing
API as well.